requirements/development.txt
-r base.txt

# Add development-specific packages
django-debug-toolbar
